Break CommandBase into many helpers.
